backend/requirements.txt
# Core
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Environment & DB helpers
python-dotenv==1.0.0
dj-database-url==1.0.0
psycopg2-binary==2.9.7

# CORS
django-cors-headers==4.0.0

# Channels (optional - used when you enable real-time chat)
channels==4.2.2
channels_redis==4.3.0

# Payments / Requests / Firebase
requests==2.31.0
firebase-admin==6.0.1
razorpay==1.4.2

# Utilities
python-multipart==0.0.6
gspread==5.12.0
google-auth==2.23.4
